+++ src/index.js
@@ -8,7 +8,7 @@ const Helmet = require( 'helmet' );
 
 const Settings = require( './settings' );
 const Cli = require( './cli' );
-const Frame = require( './frame' );
+const Html = require( './html' );
 
 // We are using express for our server
 const App = Express();
@@ -27,7 +27,7 @@ App.get( `${ Settings.endpoint }*`, ( request, response ) => {
 	}
 
 	// Generate HTML to send back to user
-	const html = Frame.GenerateHTML( request._parsedUrl.pathname, request.query, Settings.endpoint, Settings.path.templates );
+	const html = Html.GenerateHTML( request._parsedUrl.pathname, request.query, Settings.endpoint, Settings.path.templates );
 
 	// Send back the HTML to the user
 	response.send( html );
